Support adding custom events
