+++ CHANGELOG.md
@@ -1,0 +1,7 @@
+# Changelog
+
+## [Unreleased]
+## Added
+- Support adding custom events. (Check the tests/tests.rs file for an example.)
+## Changed
+## Removed
