+++ Cargo.toml
@@ -12,27 +12,32 @@ repository  = "https://github.com/GopherJ/json-rules-engine-rs"
 version     = "0.9.3"
 
 [dependencies]
-futures-util = "0.3.8"
-mustache     = "0.9.0"
-reqwest      = { version = "0.10.10", features = ["json", "rustls-tls"] }
-rhai         = { version = "0.19.7", features = [
+futures-util = { version = "0.3", optional = true }
+mustache     = "0.9"
+reqwest      = { version = "0.11", features = ["json", "rustls-tls"], optional = true }
+rhai         = { version = "0.20", features = [
   "sync",
   "f32_float",
   "no_function",
-  "no_float",
   "no_optimize",
   "no_module",
   "no_closure",
   "serde",
   "unchecked",
 ], optional = true }
-sendgrid     = { version = "0.15.0", default-features = false, features = ["async", "rustls"], optional = true }
-serde        = { version = "1.0.117", features = ["derive"] }
-serde_json   = { version = "1.0.59" }
-thiserror    = "1.0.22"
+sendgrid     = { version = "0.17", default-features = false, features = ["async", "rustls"], optional = true }
+serde        = { version = "1.0", features = ["derive"] }
+serde_json   = { version = "1.0" }
+thiserror    = "1.0"
+async-trait  = "0.1"
+
+[dev-dependencies]
+tokio = { version = "1", features = ["full"] }
+#anyhow = "1"
 
 [features]
-default = ["email"]
+default = ["eval"]
 
-email = ["sendgrid"]
+callback = ["reqwest"]
+email = ["sendgrid", "futures-util"]
 eval  = ["rhai"]
