+++ rustfmt.toml
@@ -1,1 +1,3 @@
 max_width = 80
+imports_granularity = "Crate"
+space_after_colon = true
